output files, small changes to NB, evaluate redo
All output files are properly generated. Made small changes to naive bayes (addition instead of multiplication, fixes that were already on the testbayes branch). Tried another type of logic for evaluate, it still isn't exactly what we want it to be. Perhaps it really is naive bayes causing our troubles.